Move code and rename some files
