+++ src/dq_suite/__init__.py
@@ -1,7 +1,8 @@
 """DQ API."""
 
-from src.dq_suite.df_checker import ValidationSettings, run
+from src.dq_suite.common import ValidationSettings
+from src.dq_suite.df_checker import run
 from src.dq_suite.input_helpers import schema_to_json_string
 
 # Use __all__ to let developers know what is part of the public API.
-__all__ = ["schema_to_json_string", "ValidationSettings", "run"]
+__all__ = ["schema_to_json_string", "run", "ValidationSettings"]
