+++ src/dq_suite/common.py
@@ -1,7 +1,9 @@
 from dataclasses import dataclass
 from typing import Any, Dict, List, Literal
 
-from pyspark.sql import DataFrame
+from great_expectations import get_context
+from great_expectations.data_context import AbstractDataContext
+from pyspark.sql import DataFrame, SparkSession
 
 
 @dataclass()
@@ -79,3 +81,49 @@ def write_to_unity_catalog(
     df.write.mode(mode).option("overwriteSchema", "true").saveAsTable(
         full_table_name
     )  # TODO: write as delta-table? .format("delta")
+
+
+def get_data_context(
+    data_context_root_dir: str = "/dbfs/great_expectations/",
+) -> AbstractDataContext:
+    return get_context(context_root_dir=data_context_root_dir)
+
+
+@dataclass()
+class ValidationSettings:
+    spark_session: SparkSession
+    catalog_name: str
+    table_name: str
+    check_name: str
+    data_context_root_dir: str = "/dbfs/great_expectations/"
+    data_context: AbstractDataContext | None = None
+    expectation_suite_name: str | None = None
+    checkpoint_name: str | None = None
+    run_name: str | None = None
+
+    def initialise_or_update_attributes(self):
+        self._set_data_context()
+
+        # TODO/check: do we want to allow for custom names?
+        self._set_expectation_suite_name()
+        self._set_checkpoint_name()
+        self._set_run_name()
+
+        # Finally, apply the (new) suite name to the data context
+        self.data_context.add_or_update_expectation_suite(
+            expectation_suite_name=self.expectation_suite_name
+        )
+
+    def _set_data_context(self):
+        self.data_context = get_data_context(
+            data_context_root_dir=self.data_context_root_dir
+        )
+
+    def _set_expectation_suite_name(self):
+        self.expectation_suite_name = f"{self.check_name}_expectation_suite"
+
+    def _set_checkpoint_name(self):
+        self.checkpoint_name = f"{self.check_name}_checkpoint"
+
+    def _set_run_name(self):
+        self.run_name = f"%Y%m%d-%H%M%S-{self.check_name}"
