+++ tests/test_common.py
@@ -1,9 +1,4 @@
-import pytest
-
-from src.dq_suite.df_checker import (
-    ValidationSettings,
-    read_data_quality_rules_from_json,
-)
+from src.dq_suite.common import ValidationSettings
 
 
 class TestValidationSettings:
@@ -48,21 +43,3 @@ class TestValidationSettings:
             .__getstate__()
             .values()
         )
-
-
-class TestReadDataQualityRulesFromJson:
-    dummy_file_path = "nonexistent_file_path"
-    real_file_path = "test_data/dq_rules.json"
-
-    def test_read_data_quality_rules_from_json_raises_file_not_found_error(
-        self,
-    ):
-        with pytest.raises(FileNotFoundError):
-            read_data_quality_rules_from_json(file_path=self.dummy_file_path)
-
-    def test_read_data_quality_rules_from_json_returns_json_string(self):
-        dq_rules_json_string = read_data_quality_rules_from_json(
-            file_path=self.real_file_path
-        )
-        assert type(dq_rules_json_string) is str
-        assert "tables" in dq_rules_json_string
